[fix](#32): change float -> int similar_percent

See merge request scb/crm-services!247

+++ app/api/v1/endpoints/cif/basic_information/identity/identity_document/ocr_schema_response.py
@@ -190,7 +190,7 @@ class OCRBackSideCitizenCardResponse(BaseSchema):
 # So sánh khuôn mặt đối chiếu với khuôn mặt trên giấy tờ định danh
 ########################################################################################################################
 class CompareSuccessResponse(BaseSchema):
-    similar_percent: float = Field(
+    similar_percent: int = Field(
         ...,
         description="Tỉ lệ chính xác giữa khuôn mặt đối chiếu với khuôn mặt trên giấy tờ định danh"
     )
